+++ photomosaic.py
@@ -22,16 +22,14 @@ import logging
 import time
 import random
 import numpy as np
-import scipy
-import scipy.misc
-from scipy.cluster import vq
-from scipy import interpolate
+#from scipy import interpolate
 import Image
 import ImageFilter
 import color_spaces as cs
 from progress_bar import progress_bar
 from sql_image_pool import SqlImagePool
 from tile import Tile
+from image_functions import *
 
 # Configure logger.
 FORMAT = "%(name)s.%(funcName)s:  %(message)s"
@@ -52,47 +50,6 @@ def simple(image_dir, target_filename, dimensions, output_file):
     logger.info('Saving mosaic to %s', output_file)
     mos.save(output_file)
     pool.close()
-
-def split_regions(img, split_dim):
-    """Split an image into subregions.
-    Use split_dim=2 or (2,2) or (2,3) etc.
-    Return a flat list of images."""
-    if isinstance(split_dim, int):
-        rows = columns = split_dim
-    else:
-        columns, rows = split_dim
-    r_size = img.size[0] // columns, img.size[1] // rows
-    # regions = [[None for c in range(columns)] for r in range(rows)]
-    regions = columns*rows*[None]
-    for y in range(rows):
-        for x in range(columns):
-            region = img.crop((x*r_size[0], 
-                             y*r_size[1],
-                             (x + 1)*r_size[0], 
-                             (y + 1)*r_size[1]))
-            # regions[y][x] = region ## for nested output
-            regions[y*columns + x] = region
-    return regions
-    
-def split_quadrants(img):
-    """Convenience function: calls split_regions(img, 2). Returns
-    a flat 4-element list: top-left, top-right, bottom-left, bottom-right."""
-    if img.size[0] & 1 or img.size[1] & 1:
-        logger.debug("I am quartering an image with odd dimensions.")
-    return split_regions(img, 2)
-
-def dominant_color(img, clusters=5, size=50):
-    """Group the colors in an image into like clusters, and return
-    the central value of the largest cluster -- the dominant color."""
-    assert img.mode == 'RGB', 'RGB images only!'
-    img.thumbnail((size, size))
-    imgarr = scipy.misc.fromimage(img)
-    imgarr = imgarr.reshape(scipy.product(imgarr.shape[:2]), imgarr.shape[2])
-    colors, dist = vq.kmeans(imgarr, clusters)
-    vecs, dist = vq.vq(imgarr, colors)
-    counts, bins = scipy.histogram(vecs, len(colors))
-    dominant_color = colors[counts.argmax()]
-    return map(int, dominant_color) # Avoid returning np.uint8 type.
 
 def open(target_filename):
     "Just a wrapper for Image.open from PIL"
