+++ sql_image_pool.py
@@ -4,6 +4,7 @@ import logging
 from directory_walker import DirectoryWalker
 from progress_bar import progress_bar
 import Image
+from image_functions import *
 
 # Configure logger.
 FORMAT = "%(name)s.%(funcName)s:  %(message)s"
